Accounted for categories not being provided
I set the category using the subcategory, and left the subcategory blank.
Added a unit test for this, including a doc to check and an expected result doc.
Changed how CategoryModel and RegulationModel check IsValid.
Changed DescriptionParser to log when a paragraph items type isn't accounted for.
Renamed a few poorly named variables in CategoryParser.