Adjusted Putting and request creation
Changed the Put method to have a name more fitting the fact it's not a put method anymore.
Renamed the request creation method to match the RestSharp change.
Changed how request creation was done to make it more understandable and therefore changeable.
Added overloads of the request creation method to make changes easier.