Added the bulk of it, but not quite working yet
Added a method to StorageUtilities for verification.
Added 2 methods to JiraRequestUtilites. One to get GetIssueByIssueKey for GetIssueByClauseId(). The Other to GetIssueAttachmentById for the verification.
Changed the look of the ClauseIds with extras stuff on the end, because it wasn't working with the GetIssueByClauseId for some reason. Also, renamed ClauseID to ClauseId.
Added to IssueRoot.